Game auth refactor
