+++ DragaliaAPI/DragaliaAPI/Features/Tool/FeatureExtensions.cs
@@ -11,22 +11,6 @@ public static partial class FeatureExtensions
     public static IServiceCollection AddToolFeature(this IServiceCollection serviceCollection)
     {
         serviceCollection.AddScoped<IAuthService, AuthService>();
-
-        serviceCollection
-            .AddAuthentication()
-            .AddJwtBearer(
-                AuthConstants.SchemeNames.GameJwt,
-                options =>
-                {
-                    options.Events = new()
-                    {
-                        OnMessageReceived = ToolAuthenticationHelper.OnMessageReceived,
-                        OnTokenValidated = ToolAuthenticationHelper.OnTokenValidated,
-                    };
-                    // The rest is configured in ConfigureJwtBearerOptions.cs after the ServiceProvider is built.
-                }
-            );
-
         return serviceCollection;
     }
 }
